Consolidated QC and app auth flows, removed reliance on Google+ (which is deprecated)
