+++ package.json
@@ -176,7 +176,6 @@
     "cordova-plugin-whitelist": "^1.3.3",
     "cordova-plugin-x-socialsharing": "^5.4.0",
     "fetch-mock": "^6.5.0",
-    "husky": "^1.1.4",
     "jasmine": "^3.2.0",
     "jasmine-core": "^3.2.1",
     "jasmine-expect": "^3.8.4",
